Expand the EINTR checker to a bunch more functions
PEP 0475 provides both a solution to the EINTR problem and a glimpse at
the scope of its current, awful state. Make use of this information.

Also remove os.close and os.dup2 from the list of functions to watch out
for, since it turns out that those functions should *never* be retried.
What fun!