Don't modify the locale to load translations.
This causes a crash while restoring the locale if the user has set some
of the $LC_* variables, since that causes python to return a more
complicated string that it then can't read back in. Also, the underlying
issue this was working around
(https://bitbucket.org/izi/polib/issues/54/) was fixed a while back, so
just hit delete until the problem goes away.